Ability costs are now granular

+++ src/CommandingOfficers/CommanderAbility.java
@@ -9,17 +9,17 @@ public abstract class CommanderAbility
   public static final int PHASE_TURN_END = PHASE_BUY << 1;
   public final Commander myCommander;
   protected String myName;
-  private int myPowerCost;
+  private double myPowerCost;
   public int AIFlags = PHASE_TURN_START;
 
-  public CommanderAbility(Commander commander, String abilityName, int powerCost)
+  public CommanderAbility(Commander commander, String abilityName, double powerCost)
   {
     myCommander = commander;
     myName = abilityName;
     myPowerCost = powerCost;
   }
 
-  public int getCost()
+  public double getCost()
   {
     return myPowerCost;
   }
